ContentU

+++ 14_Spring-Data-JPA/src/main/java/lk/ijse/spring/service/ItemServiceImpl.java
@@ -13,7 +13,7 @@ import java.util.ArrayList;
 
 @Service
 @Transactional
-public class ItemServiceImpl implements ItemService{
+public class ItemServiceImpl implements ItemService {
 
     @Autowired
     ItemRepo repo;
@@ -21,19 +21,32 @@ public class ItemServiceImpl implements ItemService{
     ModelMapper mapper;
 
     @Override
-    public void saveItem(ItemDTO dto){
+    public void saveItem(ItemDTO dto) {
+        if(repo.existsById(dto.getCode())){
+            throw new RuntimeException("Item "+dto.getCode()+" Already Exists!");
+        }
         repo.save(mapper.map(dto, Item.class));
     }
+
     @Override
-    public void updateItem(ItemDTO dto){
+    public void updateItem(ItemDTO dto) {
+        if(!repo.existsById(dto.getCode())){
+            throw new RuntimeException("Item "+dto.getCode()+" Does Not Exist!");
+        }
         repo.save(mapper.map(dto, Item.class));
     }
+
     @Override
-    public void deleteItem(String code){
+    public void deleteItem(String code) {
+        if(!repo.existsById(code)){
+            throw new RuntimeException("Item "+code+" Does Not Exists!");
+        }
         repo.deleteById(code);
     }
+
     @Override
-    public ArrayList<ItemDTO> getAllItems(){
-        return mapper.map(repo.findAll(), new TypeToken<ArrayList<ItemDTO>>(){}.getType());
+    public ArrayList<ItemDTO> getAllItems() {
+        return mapper.map(repo.findAll(), new TypeToken<ArrayList<ItemDTO>>() {
+        }.getType());
     }
 }
